Turnstile lock implementation, turn kernel lock into one.
BMTX stands for "blocking mutex", and it's implemented as a turnstile
exclusive recursive lock.

The Kernel Lock has been made a bmtx.

Background
~~~~~~~~~~

I've been tracking the performance impacts of the smpns changes and
the one thing who really hurt performance was turning kernel lock into
a rrwlock. The number of IPIs when building the kernel or world would
go over 5k/s, much higher than the <1k/s of the original mplock
implementation. That should be no surprise, since rrwlock is an always
sleeping lock which goes through wakeup/tsleep on every contention.

Performance also degraded considerably, baseline would build an image
in ~4m, taking ~5m25s after the change to rrwlock.

Therefore, the motivation for a turnstile lock was pure performance
wise, as they hold the same semantics as a sleeping lock like rwlock.

Semantics
~~~~~~~~~

 o Bmtx uses no critical section and doesn't block interrupts on any
way, they are resilient to preemption. They are adequate to interlock
against interrupt threads, so that should become the normal lock
throughout the kernel. This is the attempt of having synchronous
locking across interrupts and kernel threads.
 o Recursion is allowed, options for forbiding it in the future should
be discussed.
 o Interleaving on unlocking is allowed, the following is legal:

bmtx_lock(a);
bmtx_lock(b);
bmtx_unlock(a);     OR     bmtx_unlock(b);
bmtx_unlock(b);            bmtx_unlock(a);

 o Sleeping with a bmtx held is legal.

Implementation
~~~~~~~~~~~~~~

 o The cost for an uncontested operation is never higher than the
maximum cost of a single compare-and-set operation, average cost on a
cached situation is less than 50cycles per pair of lock/unlock on my
haswell i5 2.4ghz, considering function call overhead to be 0. We can
easily cut the function call overhead by making the function into a
macro that just tries the fast operation outside (one compare-and-set
call), if it fails, it calls into the function.

o Bmtx synchronizes across one atomic word called bmtx_lock, in which we
store:

From bits 2-31 or 2-63: the curproc address structure.
The first two bits are abused to store two flags which are not mutually exclusive.
BMTX_RECURSE - signals that this bmtx was recursed.
BMTX_WAITERS - signals that there are sleepers waiting for the acquisition of the lock.

o Bmtx is a turnstile lock since it tries to spin _as long_ as the
owner of the lock is currently _running_ on another cpu. Under all
other conditions, the caller blocks. Blocking means setting the
BMTX_WAITERS in the atomic word and inserting itself in the bmtx's
waitqueue.

o To be able to block on the lock, we can't depend only on the atomic
word, we need a way to interlock with the other cpu, so that we don't
miss a wakeup. This is implemented using SCHED_LOCK as an interlock,
this allows us to overcome the lost-wakeup race. The algorithm is
something like this:

1 spin on the lock until holder is not SONPROC (running).  The test
for SONPROC is done on a racy fashion, which follows the process
structure and checks p->p_stat.

2 If we think the owner appears to not be SONPROC, we interlock with
the SCHED_LOCK and _retest_ the condition, therefore, recovering from
the race.

3 Assuming the owner is indeed blocked, we set BMTX_WAITERS and put
ourselves in the queue.

On the unlock path, the fast path only attempts to clear ownership if
BMTX_WAITERS is _not_ set, if it is, we need the slower path, where we
interlock with SCHED_LOCK, saving ourselves from the race.

In the future, when the scheduler doesn't suck horse-ass, we will
interlock with an arbitrary lock, something like a stripelock, instead
of a global SCHED_LOCK. We could do it with mutexes and msleep now, but
this is so ridiculous as it's just another lock to pass ownership for
the SCHED_LOCK. A performant implementation requires the scheduler to
expose the underlying interlock to the caller.

CAVEATS
~~~~~~~

On the lock path, when testing if the owner is SONPROC, we have the
following comment:
	/* XXX if we are preempted here, high chance p might have gone and
	 * we're fucked. This is an "accepted" race in freebsd because of the
	 * way they recycle the td structures, it might not be to us. We could
	 * think of a crit_enter()/crit_leave().

We can not safely test p and deference p->p_stat, the process might have
gone through sched_exit() and p might be freed, the timing would have
to be very very very evil, I've never hit this situation and I'm
buying it for now.

Alternative solutions is changing the way we alocate proc{} (like
freebsd), or keeping track of all held mutexes and updating a state on
the mutex itself saying "OWNER_IS_RUNNING".

Results
~~~~~~~

My building times went to be as performant as baseline, sometimes
slightly faster, I could only test this in vmware so far, so I can't
have precise numbers. I would go as far as saying that it is _no_
worse than the baseline kernel.

Building time improved around 25% against the rrwlock implementation.

IPIs were reduced from 5000-6000 to 200-1200/s.

Statistics
~~~~~~~~~~

I've collected the number of fast vs slow paths by defining BMTX_STATS
on the lock, as shown below, the number of slow paths are
insignificant compared to the fast paths:

bmtx_fast_locks              21017802
bmtx_fast_unlocks            31133706
bmtx_recursed_locks          524294
bmtx_recursed_unlocks        524279
bmtx_lock_blocks             33174
bmtx_lock_blocks_cancelled   29742
bmtx_lock_blocks_slept       3862
bmtx_spun_locks              10117383
bmtx_blocked_unlocks         3866
bmtx_unlock_has_waiter       3866